update(order): add installment for order

+++ src/main/java/com/swp391/evdealersystem/dto/response/ElectricVehicleResponse.java
@@ -2,11 +2,13 @@ package com.swp391.evdealersystem.dto.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ElectricVehicleResponse {
     private Long vehicleId;
-    private Long cost;
-    private Long price;
+    private BigDecimal cost;
+    private BigDecimal price;
     private Integer batteryCapacity;
 
     private Long modelId;
